Moved save image function from commandlineinterface to imagehandler
